+++ lib/downloads.rb
@@ -22,10 +22,12 @@ namespace :download do
   task :cookbooks do
     berks_loc = Global::Settings.locations['berkshelf']
     
-    puts "\n-> running berks install --path #{HOME}/#{berks_loc} -c #{HOME}/#{berks_loc}/berks-config.json".underline    
-    puts "\nremoving Berksfile.lock...".underline
-
+    System.msgDebug "running berks install --path #{HOME}/#{berks_loc} -c #{HOME}/#{berks_loc}/berks-config.json", "system"
+    
+    System.msgDebug "removing Berksfile.lock...", "system"
     FileUtils.rm_rf "#{HOME}/#{berks_loc}/Berksfile.lock"
+
+    System.msgDebug "removing cookbooks...", "system"
     FileUtils.rm_rf "#{HOME}/#{berks_loc}/cookbooks"
 
     System.shell_cmd(
@@ -33,6 +35,8 @@ namespace :download do
       "berks install --path #{HOME}/#{berks_loc}/cookbooks -c #{HOME}/#{berks_loc}/berks-config.json",
       "Downloaded cookbooks via berkshelf to [#{berks_loc}]"
     )
+
+    System.msgSuccess "Download of cookbooks complete.", "system"
   end
 
   desc "download repos"
@@ -59,12 +63,12 @@ namespace :download do
       Dir.mkdir(download_loc) unless Dir.exists?(download_loc)
 
       if track
-        puts "-> Tracking #{full_loc}...".light_yellow
+        System.msgInfo "Tracking #{full_loc}...", "system"
         FileUtils.rm_rf "#{full_loc}"
       end
 
       if !File.exists?(full_loc)
-        puts "-> Downloading file: [#{url}] as [#{name}]".light_blue
+        System.msgInfo "Downloading file: [#{url}] as [#{name}]", "system"
 
         System.shell_cmd(
           "./",
@@ -74,6 +78,23 @@ namespace :download do
       end
     end
 
-    puts "\n== Downloads complete ==".white.on_light_blue
+    System.msgSuccess "== Downloads complete ==", "system"
+  end
+
+  desc "download plugins"
+  task :plugins do
+    def installPlug(plug_name)
+      System.shell_cmd(
+        "./",
+        "vagrant plugin install #{plug_name}",
+        "Installing vagrant plugin: [#{plug_name}]"
+      )
+    end
+
+    vagrant_plugins = Global::Settings.defaults['vagrant_plugins']
+    vagrant_plugins.each do |plug|
+      installPlug(plug)
+    end
+
   end
 end