+++ manifests/init_scripts/chef_prep.sh
@@ -9,10 +9,10 @@
 
 
 # Copy to ~/.chef/ so to make using knife easier
-if [ ! -d $CHEF_USER_HOME ]; then
-  mkdir $CHEF_USER_HOME
+if [ ! -d $HOME_CHEF_USER ]; then
+  mkdir $HOME_CHEF_USER
 fi
-cp -R /home/vagrant/manifests/chef-server/.chef $CHEF_USER_HOME
-chown -R vagrant:vagrant $CHEF_USER_HOME/.chef
+cp -R /home/vagrant/manifests/chef-server/.chef $HOME_CHEF_USER
+chown -R vagrant:vagrant $HOME_CHEF_USER/.chef
 
 echo "=> [.chef] dir copied."