+++ manifests/init_scripts/dns/init.sh
@@ -1,14 +1,16 @@
 #!/bin/bash
 
 #
+# Incudes
+# =============
+#
+. /home/vagrant/manifests/init_scripts/lib_functions.sh
+
+
+#
 # Init Base
 # =============
 #
 bash /home/vagrant/manifests/init_scripts/base_init.sh
 
 
-#
-# Incudes
-# =============
-#
-. /home/vagrant/manifests/init_scripts/lib_functions.sh